transfered todo in readme, still refactoring
