+++ readme.md
@@ -1,0 +1,19 @@
+# FEATURES
+- [ ] Normalize the movement vector for consistent speed in all directions, including diagonals.
+- [ ] Add cub3d.
+- [ ] Implement a digit scroll animation effect for the divergence meter to enhance visual feedback and user experience.
+- [ ] Store the divergence meter value in local storage to persist the state across sessions.
+- Use original Font instead of image for divergence meter.
+
+# BUGS
+
+# REFACTOR
+- [ ] Refactor the entire codebase for better readability, maintainability, and performance.
+- [ ] Change Kaboom.js to Kaplay (a maintained fork of Kaboom.js, which is deprecated)
+
+# OPTIMIZATION
+- [ ] Remove event listeners when they are no longer needed to prevent memory leaks and improve performance. (you can use  { once: true })
+
+
+# MAINTAINABILITY
+- [ ] Reorganize assets in the public folders for better structure, accessibility, and maintainability.
