+++ src/elementFactory.js
@@ -1,0 +1,108 @@
+import { SCALE_FACTOR, OFFSET_X, OFFSET_Y } from "./constants.js";
+
+export function createTile(k, tiles, frame, x, y, z) {
+    const originalSprite = k.add([
+        k.sprite(tiles, { frame }),
+        k.pos(x * SCALE_FACTOR, y * SCALE_FACTOR),
+        k.scale(SCALE_FACTOR),
+        k.z(z),
+    ]);
+
+    return originalSprite;
+}
+
+export function createInteractable(k, tiles, boundary, frame, modifX, modifY) {
+    const originalSprite = k.add([
+        k.sprite(tiles, { frame }),
+        k.pos((boundary.x + OFFSET_X + modifX) * SCALE_FACTOR, (boundary.y + OFFSET_Y + modifY) * SCALE_FACTOR),
+        k.scale(SCALE_FACTOR),
+    ]);
+
+    return {
+        originalSprite,
+        blink: false,
+    };
+}
+
+export function createIndicator(x, y, animation, k) {
+    const indicator = k.add([
+        k.sprite("indicator"),
+        k.pos((x + OFFSET_X) * SCALE_FACTOR, (y + OFFSET_Y) * SCALE_FACTOR),
+        k.scale(SCALE_FACTOR),
+    ]);
+    indicator.play(animation);
+    return indicator;
+}
+
+export function getIndicatorOffset(boundary, indicatorScale, offset) {
+    const topLeft = indicatorScale[0];
+    const topRight = indicatorScale[1];
+    const bottomLeft = indicatorScale[2];
+    const bottomRight = indicatorScale[3];
+    return [
+        {
+            dx: -offset * topLeft.x,
+            dy: -offset * topLeft.y,
+            direction: "top_left",
+        },
+        {
+            dx: -offset * bottomLeft.x,
+            dy: boundary.height - offset * bottomLeft.y,
+            direction: "bot_left",
+        },
+        {
+            dx: boundary.width - offset * topRight.x,
+            dy: -offset * topRight.y,
+            direction: "top_right",
+        },
+        {
+            dx: boundary.width - offset * bottomRight.x,
+            dy: boundary.height - offset * bottomRight.y,
+            direction: "bot_right",
+        },
+    ];
+}
+
+export function createHoverEvents(k, options) {
+    let {
+        name,
+        bubbleX,
+        bubbleY,
+        bubbleScale,
+        textSize,
+        textWidth,
+        textX,
+        textY,
+        sprite = "hover",
+        font = "myFont",
+    } = options;
+    let bubble = null;
+    let bubbleText = null;
+
+    k.onHover(name, (obj) => {
+        if (!bubble) {
+            bubble = k.add([k.sprite(sprite), k.pos(obj.pos.x + bubbleX, obj.pos.y + bubbleY), k.scale(bubbleScale)]);
+
+            bubbleText = k.add([
+                k.text(name, {
+                    size: textSize,
+                    width: textWidth,
+                    font: font,
+                }),
+                k.pos(obj.pos.x + textX, obj.pos.y + textY),
+            ]);
+        }
+    });
+
+    k.onHoverEnd(name, () => {
+        if (bubble) {
+            k.destroy(bubble);
+            bubble = null;
+        }
+
+        if (bubbleText) {
+            k.destroy(bubbleText);
+            bubbleText = null;
+        }
+    });
+}
